requirements.txt
Django==3.2.19
django-simple-history==3.0.0